FixtureByStatys Works
Now fitureByStatus with status NS retrive the data of all the match Not Started 	but Scheduled, next step will be to implement the prediction system for the fixture in NS status. Later i will implement a method to retrive all the data i need(crawler, scraper, API) to implement my personal prediction system by implementing the services for predictions.